Set a default for docs path templates
Even though the playbook for a promote job doesn't use the
afs.docs_tag_path variable, since it uses afs.docs_master_path
Ansible apparently tries to expand all the keys under afs, so they
all need to evaluate.  Supply a default for the zuul.tag component
so that happens, but set the default to "_error" so if the variable
is accidentally used in the wrong place, we don't overwrite the
branch tip docs.

Change-Id: I00f8bcde11a2a5c1f12420035be3464af0856122