Trigger gate after reverify
We don't do clean-check in zuul, which is good - it means we can
go ahead and +A something and not wait for check. However, if
something systemic has happened and a patch fails, the only way
to recheck that is for someone with +A access to re-+A something.

Add a 'reverify' comment trigger to allow someone to request a
patch to be re-run in the gate.

Change-Id: I758fcdfd50ce8c7ce6fcc5a3878fb94003c80349