Include a check queue for all projects
This patch adds every project into the check queue in the Zuul
tenant to make sure that it does zuul.yaml validation and allows
for an initial self-testing commit.

This is needed for Zuul because in openstack/project-config we
define a system-required template which does this for all the
projects.

Change-Id: I80e9f59edc23d63aac9b098eea411a9e023b19f1